Add examples in `platform_pki_2_electric_boogaloo`: Update `verify_message` example
